Refactorizadas las inserciones, ahora se realizan todas mandando un statement
